# Speedup jar package scanning