--- a/src/main/java/cpw/mods/jarhandling/SecureJarRuntime.java
+++ b/src/main/java/cpw/mods/jarhandling/SecureJarRuntime.java
@@ -1,0 +1,9 @@
+package cpw.mods.jarhandling;
+
+import java.util.Set;
+
+public interface SecureJarRuntime {
+    default Set<String> ignoredRootPackages() {
+        return Set.of("META-INF");
+    }
+}