I noted that building the module references was quite slow on the FC pack, so I propose these two simple changes:
- Some mods have a ton of assets and it takes forever to scan them... Chipped alone takes 2 seconds. So I added the infrastructure to skip some folders in the JARs. With https://github.com/neoforged/FancyModLoader/pull/14, this skips scanning the `assets` and `data` mod files for JARs. 
- Build `ModuleReference`s in parallel because why not.

Performance numbers on the FC pack (single run per line so not super precise but VisualVM confirms the observations):
```
Baseline: 4292ms
Added root package filter without FML support: 4207ms
Added root package filter with FML support: 1136ms
Additionally make jar resolution parallel: 340ms
```