@ichttt 2023-08-22
Just out of curiosity: Can you run the JarModuleFinderBenchmark? I know the set of jars is not optimal for testing the speedup of parallel scanning, but I think they still provide some value.
I like the idea of blacklisting directories that should not contain any code, but maybe we should add something in the Build Pipeline or userdev that warns if classes are present in these Directories, as it could lead to confusion why some stuff doesn't work otherwise. But tbh that is probably not needed, as it would be very weird to store classes there.

Overall, nice work!