@Technici4n 2023-08-23
Parallel scanning is only useful for jar metadatas that lazily compute their packages. The metadata implementations in SJH compute it eagerly (but `ModJarMetadata` computed it lazily and hence benefits from this speedup).

As you can see here the main hit from `JarModuleFinderBenchmark` is `getPackages`, which is not parallel with the default metadata:
![image](https://github.com/McModLauncher/securejarhandler/assets/13494793/f7707d8d-09d0-4184-889f-0cab507bc468)