@Technici4n 2023-08-25
Default jar metadata implementations should be changed too to benefit from parallelization, that might save 1s of startup in some cases.