@Technici4n 2023-11-27
Superseded by #47 and #57.